Use natural sorting to compare versions.
There is an accute issue where the 'current' link points to the wrong build. The first build has the name "r1~beta2_hrev54154_7, and the most recent build has the name r1~beta2_hrev54154_47. The current sorting algorithm ranks the first build as newer than the latest.

This also futureproofs this microservice when the hrev will increase from a 5-digit number to 6-digits (e.g. hrev99999 to hrev100000)